+++ src/python_testing/TC_TLSCERT.py
@@ -33,297 +33,26 @@
 #     factory-reset: true
 #     quiet: true
 # === END CI TEST ARGUMENTS ===
-import datetime
 import random
 import string
-from typing import List, Optional, Union
 
 import test_plan_support
-from cryptography import x509
-from cryptography.hazmat.primitives import hashes, serialization
-from cryptography.hazmat.primitives.asymmetric import ec, rsa, utils
-from cryptography.x509 import CertificateBuilder, random_serial_number
-from cryptography.x509.oid import NameOID
-from ecdsa.curves import curve_by_name
 from mobly import asserts
-from pyasn1.codec.der.decoder import decode as der_decoder
-from pyasn1.error import PyAsn1Error
-from pyasn1_modules import rfc2986, rfc5480
+from TC_TLS_Utils import TLSUtils
 
 import matter.clusters as Clusters
-from matter import ChipDeviceCtrl
-from matter.clusters.Types import Nullable, NullValue
-from matter.interaction_model import InteractionModelError, Status
+from matter.interaction_model import Status
 from matter.testing import matter_asserts
-from matter.testing.conversions import hex_from_bytes
-from matter.testing.matter_testing import (MatterBaseTest, TestStep, default_matter_test_main, has_cluster, run_if_endpoint_matches,
-                                           type_matches)
+from matter.testing.matter_testing import MatterBaseTest, TestStep, default_matter_test_main, has_cluster, run_if_endpoint_matches
 from matter.utils import CommissioningBuildingBlocks
 
 
 class TC_TLSCERT(MatterBaseTest):
-    class CertWithKey:
-        def __init__(self, cert: bytes, key: rsa.RSAPrivateKey):
-            self.cert = cert
-            self.key = key
-
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-    def get_key(self) -> rsa.RSAPrivateKey:
-        return rsa.generate_private_key(
-            public_exponent=65537, key_size=2048
-        )
-
-    def gen_cert(self, public_key: bytes | None = None) -> bytes:
-        return self.gen_cert_with_key(self.get_key(), public_key)
-
-    def gen_cert_chain(self, root: rsa.RSAPrivateKey, length: int):
-        signer = root
-        result = []
-        for i in range(length):
-            key = self.get_key()
-            cert = self.gen_cert_with_key(signer, key.public_key())
-            result.append(self.CertWithKey(cert, key))
-            signer = key
-        return result
-
-    def gen_cert_with_key(self, signer: rsa.RSAPrivateKey, public_key: bytes | None = None, subject: x509.Name | None = None) -> bytes:
-        if not public_key:
-            public_key = signer.public_key()
-
-        rand_suffix = "".join(
-            random.choices(string.ascii_letters + string.digits, k=16)
-        )
-        root_cert_subject = x509.Name(
-            [
-                x509.NameAttribute(NameOID.ORGANIZATION_NAME, "CSA"),
-                x509.NameAttribute(
-                    NameOID.COMMON_NAME, "TC_PAVS root " + rand_suffix
-                ),
-            ]
-        )
-        if not subject:
-            subject = root_cert_subject
-
-        cert = (
-            CertificateBuilder()
-            .subject_name(subject)
-            .issuer_name(root_cert_subject)
-            .public_key(public_key)
-            .serial_number(random_serial_number())
-            .not_valid_before(datetime.datetime.now(datetime.timezone.utc))
-            .not_valid_after(
-                datetime.datetime.now(datetime.timezone.utc) + datetime.timedelta(days=365.25*20)
-            )
-            .add_extension(
-                # We make it so that our root can only issue leaf certificates, no intermediate here.
-                x509.BasicConstraints(ca=True, path_length=0), critical=True
-            )
-            .add_extension(
-                x509.KeyUsage(
-                    digital_signature=True,
-                    content_commitment=False,
-                    key_encipherment=False,
-                    data_encipherment=False,
-                    key_agreement=False,
-                    key_cert_sign=True,
-                    crl_sign=True,
-                    encipher_only=False,
-                    decipher_only=False,
-                ),
-                critical=True,
-            )
-            .add_extension(
-                x509.SubjectKeyIdentifier.from_public_key(
-                    signer.public_key()
-                ),
-                critical=False,
-            )
-            .sign(signer, hashes.SHA256())
-        )
-        return cert.public_bytes(serialization.Encoding.DER)
-
-    def assert_valid_caid(self, caid):
-        asserts.assert_greater_equal(caid, 0, "Invalid CAID returned")
-        asserts.assert_less_equal(caid, 65534, "Invalid CAID returned")
-
-    def assert_valid_ccdid(self, caid):
-        asserts.assert_greater_equal(caid, 0, "Invalid CCDID returned")
-        asserts.assert_less_equal(caid, 65534, "Invalid CCDID returned")
-
-    def assert_valid_csr(self, response: Clusters.TlsCertificateManagement.Commands.ClientCSRResponse, nonce: bytes):
-        # Verify der encoded and PKCS #10 (rfc2986 is PKCS #10) - next two requirements
-        try:
-            temp, _ = der_decoder(response.csr, asn1Spec=rfc2986.CertificationRequest())
-        except PyAsn1Error:
-            asserts.fail("Unable to decode CSR - improperly formatted DER")
-        layer1 = dict(temp)
-
-        # Verify public key is 256 bytes
-        csr = x509.load_der_x509_csr(response.csr)
-        csr_pubkey = csr.public_key()
-        asserts.assert_equal(csr_pubkey.key_size, 256, "Incorrect key size")
-
-        # Verify signature algorithm is ecdsa-with-SHA256
-        signature_algorithm = dict(layer1['signatureAlgorithm'])['algorithm']
-        asserts.assert_equal(signature_algorithm, rfc5480.ecdsa_with_SHA256, "CSR specifies incorrect signature key algorithm")
-
-        # Verify signature is valid
-        asserts.assert_true(csr.is_signature_valid, "Signature is invalid")
-
-        # Verify response.nonceSignature is octet string of length 32
-        try:
-            # response.nonceSignature is an octet string if it can be converted to an int
-            int(hex_from_bytes(response.nonceSignature), 16)
-        except ValueError:
-            asserts.fail("Returned CSR nonceSignature is not an octet string")
-
-        # Verify response.nonceSignature is valid signature
-        baselen = curve_by_name("NIST256p").baselen
-        signature_raw_r = int(hex_from_bytes(response.nonceSignature[:baselen]), 16)
-        signature_raw_s = int(hex_from_bytes(response.nonceSignature[baselen:]), 16)
-
-        nonceSignature = utils.encode_dss_signature(signature_raw_r, signature_raw_s)
-        csr_pubkey.verify(signature=nonceSignature, data=nonce, signature_algorithm=ec.ECDSA(hashes.SHA256()))
-        return csr
-
-    class TargetedFabric:
-        def __init__(self, matter_test: MatterBaseTest, endpoint: Optional[int] = None,
-                     dev_ctrl: Optional[ChipDeviceCtrl.ChipDeviceController] = None,
-                     node_id: Optional[int] = None, fabric_index: Optional[int] = None):
-            self.test = matter_test
-            self.endpoint = endpoint
-            self.dev_ctrl = dev_ctrl
-            self.node_id = node_id
-            self.fabric_index = fabric_index
-
-        async def send_provision_root_command(
-                self, certificate: bytes, caid: Union[Nullable, int] = NullValue,
-                expected_status: Status = Status.Success) -> Union[Clusters.TlsCertificateManagement.Commands.ProvisionRootCertificateResponse, InteractionModelError]:
-            try:
-                result = await self.test.send_single_cmd(cmd=Clusters.TlsCertificateManagement.Commands.ProvisionRootCertificate(certificate=certificate, caid=caid),
-                                                         endpoint=self.endpoint, dev_ctrl=self.dev_ctrl, node_id=self.node_id, payloadCapability=ChipDeviceCtrl.TransportPayloadCapability.LARGE_PAYLOAD)
-
-                asserts.assert_true(type_matches(result, Clusters.TlsCertificateManagement.Commands.ProvisionRootCertificateResponse),
-                                    "Unexpected return type for ProvisionRootCertificate")
-                return result
-            except InteractionModelError as e:
-                asserts.assert_equal(e.status, expected_status, "Unexpected error returned")
-                return e
-
-        async def send_find_root_command(
-                self, caid: Union[Nullable, int] = NullValue,
-                expected_status: Status = Status.Success) -> Union[Clusters.TlsCertificateManagement.Commands.FindRootCertificateResponse, InteractionModelError]:
-            try:
-                result = await self.test.send_single_cmd(cmd=Clusters.TlsCertificateManagement.Commands.FindRootCertificate(caid=caid),
-                                                         endpoint=self.endpoint, dev_ctrl=self.dev_ctrl, node_id=self.node_id, payloadCapability=ChipDeviceCtrl.TransportPayloadCapability.LARGE_PAYLOAD)
-
-                asserts.assert_true(type_matches(result, Clusters.TlsCertificateManagement.Commands.FindRootCertificateResponse),
-                                    "Unexpected return type for FindRootCertificate")
-                return result
-            except InteractionModelError as e:
-                asserts.assert_equal(e.status, expected_status, "Unexpected error returned")
-                return e
-
-        async def send_remove_root_command(
-                self, caid: int,
-                expected_status: Status = Status.Success) -> InteractionModelError:
-            try:
-                result = await self.test.send_single_cmd(cmd=Clusters.TlsCertificateManagement.Commands.RemoveRootCertificate(caid=caid),
-                                                         endpoint=self.endpoint, dev_ctrl=self.dev_ctrl, node_id=self.node_id, payloadCapability=ChipDeviceCtrl.TransportPayloadCapability.LARGE_PAYLOAD)
-                return result
-            except InteractionModelError as e:
-                asserts.assert_equal(e.status, expected_status, "Unexpected error returned")
-                return e
-
-        async def send_csr_command(
-                self, nonce: bytes, ccdid: Union[Nullable, int] = NullValue,
-                expected_status: Status = Status.Success) -> Union[Clusters.TlsCertificateManagement.Commands.ClientCSRResponse, InteractionModelError]:
-            try:
-                result = await self.test.send_single_cmd(cmd=Clusters.TlsCertificateManagement.Commands.ClientCSR(ccdid=ccdid, nonce=nonce),
-                                                         endpoint=self.endpoint, dev_ctrl=self.dev_ctrl, node_id=self.node_id, payloadCapability=ChipDeviceCtrl.TransportPayloadCapability.LARGE_PAYLOAD)
-
-                asserts.assert_true(type_matches(result, Clusters.TlsCertificateManagement.Commands.ClientCSRResponse),
-                                    "Unexpected return type for ClientCSR")
-                return result
-            except InteractionModelError as e:
-                asserts.assert_equal(e.status, expected_status, "Unexpected error returned")
-                return e
-
-        async def send_provision_client_command(
-                self, certificate: bytes, ccdid: int, intermediates: List[bytes] = [],
-                expected_status: Status = Status.Success) -> InteractionModelError:
-            try:
-                result = await self.test.send_single_cmd(cmd=Clusters.TlsCertificateManagement.Commands.ProvisionClientCertificate(ccdid=ccdid, clientCertificate=certificate, intermediateCertificates=intermediates),
-                                                         endpoint=self.endpoint, dev_ctrl=self.dev_ctrl, node_id=self.node_id, payloadCapability=ChipDeviceCtrl.TransportPayloadCapability.LARGE_PAYLOAD)
-
-                return result
-            except InteractionModelError as e:
-                asserts.assert_equal(e.status, expected_status, "Unexpected error returned")
-                return e
-
-        async def send_find_client_command(
-                self, ccdid: Union[Nullable, int] = NullValue,
-                expected_status: Status = Status.Success) -> Union[Clusters.TlsCertificateManagement.Commands.FindClientCertificateResponse, InteractionModelError]:
-            try:
-                result = await self.test.send_single_cmd(cmd=Clusters.TlsCertificateManagement.Commands.FindClientCertificate(ccdid=ccdid),
-                                                         endpoint=self.endpoint, dev_ctrl=self.dev_ctrl, node_id=self.node_id, payloadCapability=ChipDeviceCtrl.TransportPayloadCapability.LARGE_PAYLOAD)
-
-                asserts.assert_true(type_matches(result, Clusters.TlsCertificateManagement.Commands.FindClientCertificateResponse),
-                                    "Unexpected return type for FindClientCertificate")
-                return result
-            except InteractionModelError as e:
-                asserts.assert_equal(e.status, expected_status, "Unexpected error returned")
-                return e
-
-        async def send_remove_client_command(
-                self, ccdid: int,
-                expected_status: Status = Status.Success) -> InteractionModelError:
-            try:
-                result = await self.test.send_single_cmd(cmd=Clusters.TlsCertificateManagement.Commands.RemoveClientCertificate(ccdid=ccdid),
-                                                         endpoint=self.endpoint, dev_ctrl=self.dev_ctrl, node_id=self.node_id, payloadCapability=ChipDeviceCtrl.TransportPayloadCapability.LARGE_PAYLOAD)
-                return result
-            except InteractionModelError as e:
-                asserts.assert_equal(e.status, expected_status, "Unexpected error returned")
-                return e
-
-        async def read_tls_cert_attribute(self, attribute):
-            cluster = Clusters.TlsCertificateManagement
-            return await self.test.read_single_attribute_check_success(endpoint=self.endpoint, dev_ctrl=self.dev_ctrl, node_id=self.node_id, cluster=cluster, attribute=attribute)
-
-        async def read_root_certs_attribute_as_map(self):
-            attribute_certs = await self.read_tls_cert_attribute(Clusters.TlsCertificateManagement.Attributes.ProvisionedRootCertificates)
-            matter_asserts.assert_list(attribute_certs, "Expected list")
-            found_certs = dict()
-            for cert in attribute_certs:
-                found_certs[cert.caid] = cert
-            return found_certs
-
-        async def read_client_certs_attribute_as_map(self):
-            attribute_certs = await self.read_tls_cert_attribute(Clusters.TlsCertificateManagement.Attributes.ProvisionedClientCertificates)
-            matter_asserts.assert_list(attribute_certs, "Expected list")
-            found_certs = dict()
-            for cert in attribute_certs:
-                found_certs[cert.ccdid] = cert
-            return found_certs
-
-        async def send_find_client_command_as_map(
-                self,
-                expected_status: Status = Status.Success):
-            find_response = await self.send_find_client_command(expected_status=expected_status)
-            found_certs = dict()
-            for cert in find_response.certificateDetails:
-                found_certs[cert.ccdid] = cert
-            return found_certs
-
-        async def send_remove_fabric_command(self, fabric_index: int):
-            resp = await self.test.send_single_cmd(cmd=Clusters.OperationalCredentials.Commands.RemoveFabric(fabric_index), endpoint=0, dev_ctrl=self.dev_ctrl)
-            asserts.assert_equal(
-                resp.statusCode, Clusters.OperationalCredentials.Enums.NodeOperationalCertStatusEnum.kOk)
-
     class TwoFabricData:
-        def __init__(self, cr1_cmd, cr2_cmd):
+        def __init__(self, cr1_cmd: TLSUtils, cr2_cmd: TLSUtils):
             self.cr1_cmd = cr1_cmd
             self.cr2_cmd = cr2_cmd
 
@@ -340,13 +69,13 @@ class TC_TLSCERT(MatterBaseTest):
                 f'{step_prefix}.5', "CR1 sends RemoveClientCertificate command with CCDID set to ccdidToClean[i] for all entries returned.", test_plan_support.verify_success()),
         ]
 
-    async def common_setup(self, step_prefix: string = "1") -> TargetedFabric:
+    async def common_setup(self, step_prefix: string = "1") -> TLSUtils:
         self.step(f'{step_prefix}.1')
         attributes = Clusters.TlsCertificateManagement.Attributes
         endpoint = self.get_endpoint(default=1)
 
         # Establishing CR1 controller
-        cr1_cmd = self.TargetedFabric(self, endpoint=endpoint)
+        cr1_cmd = TLSUtils(self, endpoint=endpoint)
 
         self.step(f'{step_prefix}.2')
         root_certs = await cr1_cmd.read_tls_cert_attribute(attributes.ProvisionedRootCertificates)
@@ -390,7 +119,7 @@ class TC_TLSCERT(MatterBaseTest):
         )
         fabric_index_cr2 = noc_resp.fabricIndex
 
-        cr2_cmd = self.TargetedFabric(self, endpoint=endpoint, dev_ctrl=cr2, node_id=cr2_dut_node_id, fabric_index=fabric_index_cr2)
+        cr2_cmd = TLSUtils(self, endpoint=endpoint, dev_ctrl=cr2, node_id=cr2_dut_node_id, fabric_index=fabric_index_cr2)
         return self.TwoFabricData(cr1_cmd=cr1_cmd, cr2_cmd=cr2_cmd)
 
     def pics_TC_TLSCERT_2_2(self):
@@ -455,11 +184,11 @@ class TC_TLSCERT(MatterBaseTest):
         cr2_cmd = setup_data.cr2_cmd
 
         self.step(2)
-        my_root_cert = [self.gen_cert(), self.gen_cert(), self.gen_cert()]
+        my_root_cert = [cr1_cmd.gen_cert(), cr1_cmd.gen_cert(), cr1_cmd.gen_cert()]
 
         self.step(3)
         response = await cr1_cmd.send_provision_root_command(certificate=my_root_cert[0])
-        self.assert_valid_caid(response.caid)
+        cr1_cmd.assert_valid_caid(response.caid)
         my_caid = [response.caid, None, None]
 
         self.step(4)
@@ -519,7 +248,7 @@ class TC_TLSCERT(MatterBaseTest):
 
         self.step(15)
         response = await cr2_cmd.send_provision_root_command(certificate=my_root_cert[2])
-        self.assert_valid_caid(response.caid)
+        cr2_cmd.assert_valid_caid(response.caid)
         my_caid[1] = response.caid
 
         self.step(16)
@@ -625,33 +354,33 @@ class TC_TLSCERT(MatterBaseTest):
         self.step(3)
         for i in range(2):
             response = await cr1_cmd.send_csr_command(nonce=my_nonce[i])
-            self.assert_valid_ccdid(response.ccdid)
+            cr1_cmd.assert_valid_ccdid(response.ccdid)
             my_ccdid[i] = response.ccdid
-            my_csr[i] = self.assert_valid_csr(response, my_nonce[i])
+            my_csr[i] = cr1_cmd.assert_valid_csr(response, my_nonce[i])
             if i > 1:
                 asserts.assert_not_equal(my_ccdid[i-1], my_ccdid[i], "CCDID should be unique")
 
         self.step(4)
         response = await cr2_cmd.send_csr_command(nonce=my_nonce[2])
-        self.assert_valid_ccdid(response.ccdid)
+        cr2_cmd.assert_valid_ccdid(response.ccdid)
         my_ccdid[2] = response.ccdid
-        my_csr[2] = self.assert_valid_csr(response, my_nonce[2])
+        my_csr[2] = cr2_cmd.assert_valid_csr(response, my_nonce[2])
 
         self.step(5)
         # Don't have to use the same root, but may as well
-        root = self.get_key()
-        certs_with_key_1 = self.gen_cert_chain(root, 10)
+        root = cr1_cmd.get_key()
+        certs_with_key_1 = cr1_cmd.gen_cert_chain(root, 10)
         my_intermediate_certs_1 = [x.cert for x in certs_with_key_1]
 
         self.step(6)
-        certs_with_key_2 = self.gen_cert_chain(root, 1)
+        certs_with_key_2 = cr1_cmd.gen_cert_chain(root, 1)
         my_intermediate_certs_2 = [x.cert for x in certs_with_key_2]
         signers = [root, certs_with_key_1[0].key, certs_with_key_2[0].key]
 
         self.step(7)
         my_client_cert = [None, None, None, None]
         for i in range(3):
-            my_client_cert[i] = self.gen_cert_with_key(signers[i], public_key=my_csr[i].public_key(), subject=my_csr[i].subject)
+            my_client_cert[i] = cr1_cmd.gen_cert_with_key(signers[i], public_key=my_csr[i].public_key(), subject=my_csr[i].subject)
 
         self.step(8)
         await cr1_cmd.send_provision_client_command(ccdid=my_ccdid[0], certificate=my_client_cert[0])
@@ -706,10 +435,10 @@ class TC_TLSCERT(MatterBaseTest):
 
         self.step(16)
         response = await cr1_cmd.send_csr_command(ccdid=my_ccdid[0], nonce=my_nonce[3])
-        self.assert_valid_ccdid(response.ccdid)
+        cr1_cmd.assert_valid_ccdid(response.ccdid)
         asserts.assert_equal(response.ccdid, my_ccdid[0], "Expected same ID")
-        my_csr[3] = self.assert_valid_csr(response, my_nonce[3])
-        my_client_cert[3] = self.gen_cert_with_key(root, public_key=my_csr[3].public_key(), subject=my_csr[3].subject)
+        my_csr[3] = cr1_cmd.assert_valid_csr(response, my_nonce[3])
+        my_client_cert[3] = cr1_cmd.gen_cert_with_key(root, public_key=my_csr[3].public_key(), subject=my_csr[3].subject)
 
         self.step(17)
         await cr1_cmd.send_provision_client_command(ccdid=my_ccdid[0], certificate=my_client_cert[3])
