+++ src/python_testing/TC_TLSCLIENT_1_1.py
@@ -33,17 +33,12 @@
 #     factory-reset: true
 #     quiet: true
 # === END CI TEST ARGUMENTS ===
-from typing import Union
-
 from mobly import asserts
+from TC_TLS_Utils import TLSUtils
 
 import matter.clusters as Clusters
-from matter import ChipDeviceCtrl
-from matter.clusters.Types import Nullable, NullValue
-from matter.interaction_model import InteractionModelError, Status
-from matter.testing.matter_testing import (MatterBaseTest, TestStep, default_matter_test_main, has_cluster, run_if_endpoint_matches,
-                                           type_matches)
-from matter.tlv import uint
+from matter.interaction_model import Status
+from matter.testing.matter_testing import MatterBaseTest, TestStep, default_matter_test_main, has_cluster, run_if_endpoint_matches
 
 
 class TC_TLSCLIENT_1_1(MatterBaseTest):
@@ -58,28 +53,16 @@ class TC_TLSCLIENT_1_1(MatterBaseTest):
         ]
         return steps
 
-    async def send_provision_command(self, endpoint: int, hostname: bytes, port: uint, caid: uint,
-                                     ccdid: Union[Nullable, uint] = NullValue, expected_status: Status = Status.Success) -> Union[Clusters.TlsClientManagement.Commands.ProvisionEndpointResponse, InteractionModelError]:
-        try:
-            result = await self.send_single_cmd(cmd=Clusters.TlsClientManagement.Commands.ProvisionEndpoint(hostname=hostname, port=port, caid=caid, ccdid=ccdid),
-                                                endpoint=endpoint, payloadCapability=ChipDeviceCtrl.TransportPayloadCapability.LARGE_PAYLOAD)
-
-            asserts.assert_true(type_matches(result, Clusters.TlsClientManagement.Commands.ProvisionEndpointResponse),
-                                "Unexpected return type for ProvisionEndpoint")
-            return result
-        except InteractionModelError as e:
-            asserts.assert_equal(e.status, expected_status, "Unexpected error returned")
-            return e
-
     @run_if_endpoint_matches(has_cluster(Clusters.TlsClientManagement))
     async def test_TC_TLSCLIENT_1_1(self):
 
         endpoint = self.get_endpoint(default=1)
+        util = TLSUtils(self, endpoint=endpoint)
 
         self.step(1)
 
         self.step(2)
-        e = await self.send_provision_command(endpoint=endpoint, expected_status=Status.Failure, hostname=b"my_hostname", port=1000, caid=10)
+        e = await util.send_provision_tls_endpoint_command(expected_status=Status.Failure, hostname=b"my_hostname", port=1000, caid=10)
         asserts.assert_equal(
             e.clusterStatus, Clusters.TlsClientManagement.Enums.StatusCodeEnum.kRootCertificateNotFound, "Unexpected error returned")
